Improve functionality around Network Controller and Service Fabric cmdlets (#265)
# Description
This pull request primarily includes changes to the `SdnDiag` module in
the `src` directory. The changes focus on modifying the way Network
Controller information is retrieved and used. The most important changes
include modifying the Network Controller's default value to use the
environment variable `COMPUTERNAME`, updating several functions to use
the first element of the `NetworkController` array, and adding a new
parameter `Name` to some functions to specify the friendly name of the
node for the network controller.

Changes to Network Controller default value:

*
[`src/modules/SdnDiag.Health/public/Debug-SdnFabricInfrastructure.ps1`](diffhunk://#diff-324c08d6d20cb9d37d1a5b9756303f40024eb3d5c4008d8b210c8e7295ea1151L25-R25):
Modified the default value of `NetworkController` to use the
`COMPUTERNAME` environment variable instead of `HostName`.

Changes to Network Controller usage:

* Various files in `src/modules/SdnDiag.Health/private/`: Updated
several functions to use the first element of the `NetworkController`
array instead of the whole array. This includes
`Test-NcHostAgentConnectionToApiService`, `Test-NcUrlNameResolution`,
`Test-ServiceFabricApplicationHealth`,
`Test-ServiceFabricClusterHealth`, `Test-ServiceFabricNodeStatus`, and
`Test-ServiceFabricPartitionDatabaseSize`.
[[1]](diffhunk://#diff-68dbaf962eab89a5e9ffd87f730d4eb00c609721954c0572f8e470a82949546aL45-R45)
[[2]](diffhunk://#diff-53875d037e2095362d1f4d2ebce09ba56c67e194efbb31a4155b7fc6a28b3a43L24-R24)
[[3]](diffhunk://#diff-668704dd9169d77f54a7cd6d4b510ec1e6dc285a592c37ad62d909547c7b99fbL23-R28)
[[4]](diffhunk://#diff-e5e0d69262c7963f0a6c15a8219173e2212dadb005c38dde67c19939b33f2f4aL23-R28)
[[5]](diffhunk://#diff-48f1eadb283a67fe794b6b4010a03fcf8d1951cb59507670bbe5dc6c2e882fc9L23-R23)
[[6]](diffhunk://#diff-bd764ff1ecb6ac1be4d69275151ea1ef4733b8114c22aadedcd95075feec9f94L24-R34)

Addition of `Name` parameter:

*
[`src/modules/SdnDiag.NetworkController/private/Get-NetworkControllerNodeInfoFromClusterManifest.ps1`](diffhunk://#diff-37ef7e80318339c547eb7ce0e854e2fc041c2d614d34761b21a304070da01e82R6-R21):
Added the `Name` parameter to specify the friendly name of the node for
the network controller. If provided, the function returns settings for
the specified node.
[[1]](diffhunk://#diff-37ef7e80318339c547eb7ce0e854e2fc041c2d614d34761b21a304070da01e82R6-R21)
[[2]](diffhunk://#diff-37ef7e80318339c547eb7ce0e854e2fc041c2d614d34761b21a304070da01e82R48-R51)
*
[`src/modules/SdnDiag.NetworkController/private/Get-SdnNetworkControllerInfoFromClusterManifest.ps1`](diffhunk://#diff-55745d83e94f36597ab0e1d970811f92239e65fc5a2185b5ee1826e3e8c783b9R7-R8):
Added the `Name` parameter to specify the friendly name of the node for
the network controller. If provided, the function returns settings for
the specified node.
[[1]](diffhunk://#diff-55745d83e94f36597ab0e1d970811f92239e65fc5a2185b5ee1826e3e8c783b9R7-R8)
[[2]](diffhunk://#diff-55745d83e94f36597ab0e1d970811f92239e65fc5a2185b5ee1826e3e8c783b9R18-R51)
*
[`src/modules/SdnDiag.NetworkController/private/Invoke-CertRotateCommand.ps1`](diffhunk://#diff-f70f348636178625815e88bf5154b7a85247a0f17b3d8a020249469d90dbc870R9-R11):
Added the `Name` parameter to specify the friendly name of the node for
the network controller. If provided, the function operates on the
specified node.
[[1]](diffhunk://#diff-f70f348636178625815e88bf5154b7a85247a0f17b3d8a020249469d90dbc870R9-R11)
[[2]](diffhunk://#diff-f70f348636178625815e88bf5154b7a85247a0f17b3d8a020249469d90dbc870L63-R66)

Other changes:

*
[`src/SdnDiagnostics.psd1`](diffhunk://#diff-17aaaa968cc894449c79b449c228b28d8a8990bde4000e59bcf24d8189671ee1R99):
Added `Get-SdnServiceFabricApplication` to the list of functions.
*
[`src/modules/SdnDiag.NetworkController/public/Get-SdnInfrastructureInfo.ps1`](diffhunk://#diff-5eb3ff3eb27aac8cf3bba7ddb23507bfa1d65b2ff6ae6ab2a176f4d871ba84e6L24-R24):
Modified the function to confirm if the current machine is a Network
Controller.
[[1]](diffhunk://#diff-5eb3ff3eb27aac8cf3bba7ddb23507bfa1d65b2ff6ae6ab2a176f4d871ba84e6L24-R24)
[[2]](diffhunk://#diff-5eb3ff3eb27aac8cf3bba7ddb23507bfa1d65b2ff6ae6ab2a176f4d871ba84e6L49-R50)

# Change type
- [x] Bug fix (non-breaking change)
- [ ] Code style update (formatting, local variables)
- [ ] New Feature (non-breaking change that adds new functionality
without impacting existing)
- [x] Breaking change (fix or feature that may cause functionality
impact)
- [ ] Other

# Checklist:
- [x] My code follows the style and contribution guidelines of this
project.
- [x] I have tested and validated my code changes.