Disable certain functions from executing on SFNC clusters (#289)
This pull request primarily focuses on enhancing the support for Service
Fabric clusters in the SdnDiagnostics module. The changes involve adding
conditional checks in various functions to ensure they are only executed
on Service Fabric clusters. Additionally, some code has been refactored
for better organization and readability.

The most significant changes are as follows:

**Added checks for Service Fabric cluster configuration:**

* `Start-SdnCertificateRotation` in `SdnDiagnostics.psm1`: Added a
condition to check if the cluster configuration type is Service Fabric.
If not, an exception is thrown. Also, the role for getting the
SdnModuleConfiguration has been changed to 'NetworkController_SF'.
* `Repair-SdnDiagnosticsScheduledTask` in
`Repair-SdnDiagnosticsScheduledTask.ps1`: Added a switch statement to
set the task name based on the cluster configuration type.
* Several functions in `SdnDiag.NetworkController.SF` and
`SdnDiag.NetworkController` modules: Added a condition to check if the
cluster configuration type is Service Fabric. If not, an exception is
thrown.
[[1]](diffhunk://#diff-324c08d6d20cb9d37d1a5b9756303f40024eb3d5c4008d8b210c8e7295ea1151R47-R50)
[[2]](diffhunk://#diff-9e1adf0cfe8d7b8cf607e163c0a96ffc5c307137717e02c2c6b241176d47b502R30-R33)
[[3]](diffhunk://#diff-da56c5aff5c8d4ae1f24d32ba504c7708faf180de322fca8cd422a318f6ca088R34-R37)
[[4]](diffhunk://#diff-de7dc7282f0b7233ad4ce1ccf8882c14469a3cb33837707792c5d660dd91605aR48-R51)
[[5]](diffhunk://#diff-67eb6850f4f3aa92b6398560376b7e8ae52e5be6b6b9504ba882588c7843e4fbL36-R37)
[[6]](diffhunk://#diff-dd2d52a041fb0b98591e81eb4c4f2a8a1807f54f1510401f498ef4e6d1451b4bR38-R41)
[[7]](diffhunk://#diff-705c2c643df865da51dd5c96aab0f2b1b892dba826f9e62e51302f6f3fd90ee1R17-R20)
[[8]](diffhunk://#diff-a54d2dff16ad518ff9e95173bfb4bdc38c7f5241c46c76e3f79f36359b672b31R17-R20)
[[9]](diffhunk://#diff-99617b1bd78044ec20af4d249f5641df0de3a0beae707ac9f9cbf97084b25278R38-R41)
[[10]](diffhunk://#diff-0b23655b2855a4a49fd3072131d78f5feec01254c422213affdc0b22b2c27eefR33-R36)

**Code refactoring:**

* `Set-SdnNetworkController` in `Set-SdnNetworkController.ps1`: The code
to set the parameters for the function has been moved after the check
for Service Fabric cluster configuration.
* `Get-SdnServiceFabricNode` in `Get-SdnServiceFabricNode.ps1`: Removed
an unnecessary line of code.

These changes aim to improve the module's functionality and
compatibility with Service Fabric clusters, ensuring that the
appropriate functions are executed based on the cluster configuration
type.