Enable native support for Failover Cluster NC (#287)
This pull request includes changes to enhance the functionality of the
SDN Diagnostics module, particularly for Service Fabric and Failover
Cluster configurations. The most significant changes are the addition of
new commands, modifications to existing commands, and the introduction
of new modules for different cluster configurations.

Additions and modifications to commands:

*
[`src/SdnDiagnostics.Helper.psm1`](diffhunk://#diff-c298d582fd8e12a03b0e7b1dc9c562c58cc978d2e9044757d73f1a00f840c6b5L72-R78):
Removed 'Test-SdnKnownIssue' and added new commands to the
`$networkControllerParamCommands` array.
*
[`src/modules/SdnDiag.Common/private/Export-RegistryKeyConfigDetails.ps1`](diffhunk://#diff-2bc0c28b695611157ad4de8a232e841315642b136775d2bf32f8f74acd670b19L18-R21):
Changed the logging level and improved the registry key path
enumeration. Added a security measure to redact the AESKey for cluster
network controller registry keys.
[[1]](diffhunk://#diff-2bc0c28b695611157ad4de8a232e841315642b136775d2bf32f8f74acd670b19L18-R21)
[[2]](diffhunk://#diff-2bc0c28b695611157ad4de8a232e841315642b136775d2bf32f8f74acd670b19R30-R34)
*
[`src/modules/SdnDiag.Common/private/Get-CommonConfigState.ps1`](diffhunk://#diff-835b08b8e1c0bce0a59f0c589152e1a7e9a6e24bfb443ad00669c63f3924d881L58-R58):
Modified the `Get-NetAdapterAdvancedProperty` command to handle errors
according to the `$ErrorActionPreference` variable.
*
[`src/modules/SdnDiag.Common/public/Repair-SdnDiagnosticsScheduledTask.ps1`](diffhunk://#diff-a69fd2713e1d24fef7f12349ac339925ef11a4e9a80269f3c5abe39f02632788R10-R18):
Added a switch statement to handle different `ClusterConfigType` values.
*
[`src/modules/SdnDiag.Health/private/Test-NcUrlNameResolution.ps1`](diffhunk://#diff-53875d037e2095362d1f4d2ebce09ba56c67e194efbb31a4155b7fc6a28b3a43L30-R32):
Changed the `Get-SdnNetworkController` command to
`Get-SdnNetworkControllerSF`.
*
[`src/modules/SdnDiag.Health/public/Debug-SdnFabricInfrastructure.ps1`](diffhunk://#diff-324c08d6d20cb9d37d1a5b9756303f40024eb3d5c4008d8b210c8e7295ea1151R47-R50):
Added a condition to throw an exception if the `ClusterConfigType` is
not 'ServiceFabric'.

New modules for different cluster configurations:

*
[`src/modules/SdnDiag.NetworkController.FC/SdnDiag.NetworkController.FC.Config.psd1`](diffhunk://#diff-656db7b7942a5c8bebb655094f4e112758a27267910335e9c06d6575c280b295R1-R70):
Created a new configuration file for the NetworkControllerFC module.
*
[`src/modules/SdnDiag.NetworkController.FC/SdnDiag.NetworkController.FC.psm1`](diffhunk://#diff-141e8a7963d2aa6f51b5851094e11e0287b3c8793904094c9840ed25816c2a2aR1-R13):
Created a new module file for the NetworkControllerFC module.
*
[`src/modules/SdnDiag.NetworkController.FC/private/Get-NetworkControllerFCConfigState.ps1`](diffhunk://#diff-a101ed5e85ead9ca9371fad33a94311da92c52d3812fd2f1b23fbb6ca7af6ee9R1-R43):
Created a new function to output a set of configuration state files for
the network controller role.
*
[`src/modules/SdnDiag.NetworkController.FC/private/Get-SdnClusterName.ps1`](diffhunk://#diff-5c8dc2849956c598e30fbe1008e22689493be806df2cf5f403958cdb2e86d99cR1-R27):
Created a new function to get the cluster name.
*
[`src/modules/SdnDiag.NetworkController.FC/private/Get-SdnNetworkControllerFC.ps1`](diffhunk://#diff-a2156b35c839b8eaf62a5ada1c1a34d7bd04c2fdf44f317d4231ef9162ec610fR1-R45):
Created a new function to get network controller application settings.
*
[`src/modules/SdnDiag.NetworkController.FC/private/Get-SdnNetworkControllerFCClusterInfo.ps1`](diffhunk://#diff-45acf2c028d64150f072773cb338f9ba97cec5a221c3679a15546e47281ce796R1-R63):
Created a new function to gather the Network Controller cluster wide
info.
*
[`src/modules/SdnDiag.NetworkController.FC/private/Get-SdnNetworkControllerFCNode.ps1`](diffhunk://#diff-07e5e85f95431cfe04c99dc1bc32352f6850f798d97dcac0cb582d4d193b47e0R1-R85):
Created a new function to return a list of servers from network
controller.
*
[`src/modules/SdnDiag.NetworkController.SF/SdnDiag.NetworkController.SF.Config.psd1`](diffhunk://#diff-fcf4a3d34ad27665c7140f0762b97036cdc06820518e3ef8d5cf04ea80279c22R1-R34):
Created a new configuration file for the NetworkControllerSF module.

File additions:

*
[`src/SdnDiagnostics.psd1`](diffhunk://#diff-17aaaa968cc894449c79b449c228b28d8a8990bde4000e59bcf24d8189671ee1R38-R39):
Added new module files to the list of files.