Misc fixes 20241007 (#325)
# Description
This pull request includes several changes to improve the error handling
and structure of the PowerShell scripts within the `SdnDiag` module. The
most important changes involve adding error handling for directory
creation, restructuring the process flow for exporting objects to files,
and adjusting the module seeding logic.

### Error Handling Improvements:

*
[`src/modules/SdnDiag.Utilities/private/Export-ObjectToFile.ps1`](diffhunk://#diff-0e551b186879f33ef93255e03c2c40a9a7ca7cea8b9ab5a6a5296b5523f886acL35-R65):
Added a `try-catch` block to handle exceptions when creating
directories, ensuring that any issues are properly caught and reported.

### Process Flow Adjustments:

*
[`src/modules/SdnDiag.Utilities/private/Export-ObjectToFile.ps1`](diffhunk://#diff-0e551b186879f33ef93255e03c2c40a9a7ca7cea8b9ab5a6a5296b5523f886acL35-R65):
Reorganized the script to check for null objects in the `process` block
rather than the `begin` block, improving the logical flow of the script.
*
[`src/modules/SdnDiag.Utilities/public/Install-SdnDiagnostics.ps1`](diffhunk://#diff-34bdcb1b67d8c832eb7f198e206f31b08d82bd5c5e66847e9a96914fe38e1550L69-R75):
Moved the check for `DisableModuleSeeding` from the `begin` block to the
`process` block, ensuring that the seeding logic is evaluated for each
computer in the `ComputerName` list.

### Code Enhancements:

*
[`src/modules/SdnDiag.Health/private/Test-ResourceConfigurationState.ps1`](diffhunk://#diff-62a5aba7f08f0a50c099f8d938e4bfcafa353a3b173e467c36f7a7ccc30cf097R79):
Added a check to handle cases where `detailedInfo` is present, ensuring
that the detailed information is processed correctly.
*
[`src/modules/SdnDiag.Health/private/Test-ResourceConfigurationState.ps1`](diffhunk://#diff-62a5aba7f08f0a50c099f8d938e4bfcafa353a3b173e467c36f7a7ccc30cf097L102-R111):
Modified the message formatting logic to handle cases where
`errorMessages` is empty, ensuring a cleaner output.

# Change type
- [x] Bug fix (non-breaking change)
- [ ] Code style update (formatting, local variables)
- [ ] New Feature (non-breaking change that adds new functionality
without impacting existing)
- [ ] Breaking change (fix or feature that may cause functionality
impact)
- [ ] Other

# Checklist:
- [x] My code follows the style and contribution guidelines of this
project.
- [x] I have tested and validated my code changes.