Fix issues with certificate rotation (#347)
# Description
Summary of changes:
- Fixed an issue when using `-CertPath` to specify directory, we do not
properly identify the .cer file, resulting in us copying the .pfx file
to SB nodes, which require a password that we are not passing.
- Change where we catch an exception in attempt to retry if generic
error is thrown.
- Fixed an issue where multiple certificates with same subject are
returned resulting in generation of new nc node certificate was failing
due to mismatched types
- Fixed an issue where we were not passing `CN=` to `Get-SdnCertificate`
resulting in cert rotation failing for environments that were using
FindBySubjectName

# Change type
- [x] Bug fix (non-breaking change)
- [ ] Code style update (formatting, local variables)
- [ ] New Feature (non-breaking change that adds new functionality
without impacting existing)
- [ ] Breaking change (fix or feature that may cause functionality
impact)
- [ ] Other

# Checklist:
- [x] My code follows the style and contribution guidelines of this
project.
- [x] I have tested and validated my code changes.