Fix integration issues with Azure Local (#370)
# Description
Summary of changes:
This pull request introduces several new functions and improvements to
the `SdnDiagnostics` module. The key changes include adding a new
function to retrieve the product name from the registry, enhancing the
verbosity of logging, and modifying how network controller commands are
invoked.

### New Functions and Enhancements:

*
[`src/SdnDiagnostics.psm1`](diffhunk://#diff-490865628c61b2e97c50f45b37d7086647c70b2444cbfb9c60cc8c682801356eR43-R60):
Added a new global configuration mode based on the product name
retrieved from the registry and included logic to handle Azure Stack HCI
environments.
*
[`src/modules/SdnDiag.Utilities.psm1`](diffhunk://#diff-9e50bcf150b088e2c4df3d0768d8dcad4abf1338de3498749997682448a07bdcR2642-R2653):
Introduced the `Get-ProductNameFromRegistry` function to fetch the
product name from the registry and determine the appropriate
configuration mode.
*
[`src/modules/SdnDiag.NetworkController.FC.psm1`](diffhunk://#diff-9d7dd8ad1b4e123a03f6440ee1a241e1831b43f5a5f981f98d018728966cfc48R289-R296):
Added the `Confirm-IsFailoverClusterNC` function to verify if the
environment is a failover cluster.

### Logging Improvements:

*
[`src/modules/SdnDiag.Common.psm1`](diffhunk://#diff-9ab71f66f6e21719dc9527f01ea738656003bbbe631f4f1bd85ab1ab8a746f24L445-R445):
Enhanced the `Export-RegistryKeyConfigDetails` function to include
verbose logging for better traceability.

### Command Invocation Modifications:

*
[`src/modules/SdnDiag.NetworkController.FC.psm1`](diffhunk://#diff-9d7dd8ad1b4e123a03f6440ee1a241e1831b43f5a5f981f98d018728966cfc48L122-R128):
Simplified the invocation of `Get-NetworkControllerOnFailoverCluster` by
directly calling the function instead of using a script block.
*
[`src/modules/SdnDiag.Server.psm1`](diffhunk://#diff-11217f20b55d3b4ea34c8c217794c81d65acc4852dff9bf4295e5cc4d6dfaeedL721-R721):
Suppressed warning output when invoking commands in the
`Get-ServerConfigState` function to reduce noise in the logs.

# Change type
- [ ] Bug fix (non-breaking change)
- [ ] Code style update (formatting, local variables)
- [x] New Feature (non-breaking change that adds new functionality
without impacting existing)
- [ ] Breaking change (fix or feature that may cause functionality
impact)
- [ ] Other

# Checklist:
- [x] My code follows the style and contribution guidelines of this
project.
- [x] I have tested and validated my code changes.