Add health test to check for multiple switches that have VFP extension enabled (#393)
# Description
Summary of changes:
This pull request primarily introduces new functionality related to
Virtual Filtering Platform (VFP) enabled virtual switches in the SDN
diagnostics module. The most significant changes include the addition of
a new cmdlet, updates to existing cmdlets, and enhancements to the
diagnostics configuration.

### New Functionality:
* Added `Get-SdnVMSwitch` cmdlet to retrieve virtual switches from the
hypervisor with an option to filter by VFP enabled switches.
* Introduced `Test-SdnVfpEnabledVMSwitch` cmdlet to check that only one
VFP enabled virtual switch is configured on the system.

### Enhancements to Existing Cmdlets:
* Updated `Get-SdnNetAdapterEncapOverheadConfig` to support filtering by
switch name or ID and to use the new `Get-SdnVMSwitch` cmdlet for
retrieving VFP enabled switches.
[[1]](diffhunk://#diff-11217f20b55d3b4ea34c8c217794c81d65acc4852dff9bf4295e5cc4d6dfaeedR1460-L1475)
[[2]](diffhunk://#diff-11217f20b55d3b4ea34c8c217794c81d65acc4852dff9bf4295e5cc4d6dfaeedR1508)

### Configuration and Diagnostics Updates:
* Added new entries for `Get-SdnVMSwitch` and
`Test-SdnVfpEnabledVMSwitch` in the `src/SdnDiagnostics.psd1` file.
[[1]](diffhunk://#diff-17aaaa968cc894449c79b449c228b28d8a8990bde4000e59bcf24d8189671ee1R119)
[[2]](diffhunk://#diff-17aaaa968cc894449c79b449c228b28d8a8990bde4000e59bcf24d8189671ee1R177)
* Included `Test-SdnVfpEnabledVMSwitch` in the diagnostics configuration
file with description and impact details.
* Integrated `Test-SdnVfpEnabledVMSwitch` into the `Debug-SdnServer`
function to be part of the diagnostic tests.

# Change type
- [ ] Bug fix (non-breaking change)
- [ ] Code style update (formatting, local variables)
- [x] New Feature (non-breaking change that adds new functionality
without impacting existing)
- [ ] Breaking change (fix or feature that may cause functionality
impact)
- [ ] Other

# Checklist:
- [x] My code follows the style and contribution guidelines of this
project.
- [x] I have tested and validated my code changes.