Fix diagnostics health test (#419)
# Description
This pull request includes changes to the
`src/modules/SdnDiag.Health.psm1` file to update the property names used
in health check functions. The most important changes include modifying
the connection state and node status checks to use the correct property
names.

Updates to property names:

*
[`src/modules/SdnDiag.Health.psm1`](diffhunk://#diff-15898640fc68e07afa836ad8d93af4f22a4442978d9c233f39d48d44d85cfb60L2357-R2357):
In the `Test-SdnHostAgentConnectionStateToApiService` function, changed
the property checked from `ConnectionState` to `State` to correctly
determine if the connection is established.
*
[`src/modules/SdnDiag.Health.psm1`](diffhunk://#diff-15898640fc68e07afa836ad8d93af4f22a4442978d9c233f39d48d44d85cfb60L2478-R2478):
In the `Test-SdnServiceFabricNodeStatus` function, changed the property
checked from `NodeStatus` to `Status` to correctly determine if the node
is up.

# Change type
- [x] Bug fix (non-breaking change)
- [ ] Code style update (formatting, local variables)
- [ ] New Feature (non-breaking change that adds new functionality
without impacting existing)
- [ ] Breaking change (fix or feature that may cause functionality
impact)
- [ ] Other

# Checklist:
- [x] My code follows the style and contribution guidelines of this
project.
- [x] I have tested and validated my code changes.